scripts/requirements.txt
altgraph==0.17.4
chardet==5.2.0
et-xmlfile==1.1.0
icalendar==6.0.1
lxml==5.3.0
macholib==1.16.3
modulegraph==0.19.6
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
py2app==0.28.8
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
